# apps/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Koyomi Gal 后端 API",
        "title": "Koyomi Gal API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/galgames/{id}/favorite": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "收藏当前用户与 Galgame 的关系，并原子更新收藏计数",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "galgames"
                ],
                "summary": "收藏 Galgame",
                "operationId": "addGalgameFavorite",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Galgame ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "收藏结果",
                        "schema": {
                            "$ref": "#/definitions/dto.FavoriteDataResponse"
                        }
                    },
                    "400": {
                        "description": "Galgame ID 格式不正确",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "用户登录失效",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Galgame 不存在",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "已收藏该 Galgame",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "收藏失败",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除当前用户与 Galgame 的收藏关系，并原子更新收藏计数",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "galgames"
                ],
                "summary": "取消收藏 Galgame",
                "operationId": "removeGalgameFavorite",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Galgame ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "收藏已取消",
                        "schema": {
                            "$ref": "#/definitions/response.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Galgame ID 格式不正确",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "用户登录失效",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Galgame 不存在或未收藏",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "取消收藏失败",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/galgames/{id}/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "返回当前用户对 Galgame 的评分、收藏和游玩状态",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "galgames"
                ],
                "summary": "查看当前用户与 Galgame 的关系",
                "operationId": "getMyGalgameRelation",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Galgame ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "用户关系详情",
                        "schema": {
                            "$ref": "#/definitions/dto.GalgameUserRelationResponse"
                        }
                    },
                    "400": {
                        "description": "Galgame ID 格式不正确",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "用户登录失效",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Galgame 不存在",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "查询用户关系失败",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/galgames/{id}/rating": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建或更新当前用户对 Galgame 的评分，并重新计算评分统计",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "galgames"
                ],
                "summary": "评分 Galgame",
                "operationId": "upsertGalgameRating",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Galgame ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "评分请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpsertRatingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "评分结果",
                        "schema": {
                            "$ref": "#/definitions/dto.RatingDataResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数格式不正确",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "用户登录失效",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Galgame 不存在",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "评分失败",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除当前用户对 Galgame 的评分，并重新计算评分统计",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "galgames"
                ],
                "summary": "删除 Galgame 评分",
                "operationId": "deleteGalgameRating",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Galgame ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "评分已删除",
                        "schema": {
                            "$ref": "#/definitions/response.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Galgame ID 格式不正确",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "用户登录失效",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Galgame 不存在或未评分",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "删除评分失败",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/galgames/{id}/state": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建或更新当前用户对 Galgame 的游玩状态和游玩时长",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "galgames"
                ],
                "summary": "设置 Galgame 游玩状态",
                "operationId": "upsertGalgameUserState",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Galgame ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "游玩状态请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpsertUserStateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "游玩状态结果",
                        "schema": {
                            "$ref": "#/definitions/dto.UserStateDataResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数格式不正确",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "用户登录失效",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Galgame 不存在",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "设置游玩状态失败",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除当前用户对 Galgame 的游玩状态",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "galgames"
                ],
                "summary": "删除 Galgame 游玩状态",
                "operationId": "deleteGalgameUserState",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Galgame ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "游玩状态已删除",
                        "schema": {
                            "$ref": "#/definitions/response.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Galgame ID 格式不正确",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "用户登录失效",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Galgame 不存在或未设置游玩状态",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "删除游玩状态失败",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.FavoriteData": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "favorited": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "dto.FavoriteDataResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "$ref": "#/definitions/dto.FavoriteData"
                },
                "msg": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "dto.GalgameUserRelationData": {
            "type": "object",
            "properties": {
                "favorite": {
                    "$ref": "#/definitions/dto.FavoriteData"
                },
                "galgame_id": {
                    "type": "integer",
                    "example": 1
                },
                "rating": {
                    "$ref": "#/definitions/dto.RatingData"
                },
                "state": {
                    "$ref": "#/definitions/dto.UserStateData"
                }
            }
        },
        "dto.GalgameUserRelationResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "$ref": "#/definitions/dto.GalgameUserRelationData"
                },
                "msg": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "dto.RatingData": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "score": {
                    "type": "integer",
                    "example": 8
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.RatingDataResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "$ref": "#/definitions/dto.RatingData"
                },
                "msg": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "dto.UpsertRatingRequest": {
            "type": "object",
            "required": [
                "score"
            ],
            "properties": {
                "score": {
                    "type": "integer",
                    "enum": [
                        1,
                        2,
                        3,
                        4,
                        5,
                        6,
                        7,
                        8,
                        9,
                        10
                    ],
                    "example": 8
                }
            }
        },
        "dto.UpsertUserStateRequest": {
            "type": "object",
            "required": [
                "state"
            ],
            "properties": {
                "play_time_minutes": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 120
                },
                "state": {
                    "type": "integer",
                    "enum": [
                        1,
                        2,
                        3,
                        4,
                        5
                    ],
                    "example": 2
                }
            }
        },
...
        "dto.UserStateData": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "play_time_minutes": {
                    "type": "integer",
                    "example": 120
                },
                "state": {
                    "type": "integer",
                    "example": 2
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.UserStateDataResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "$ref": "#/definitions/dto.UserStateData"
                },
                "msg": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 233
                },
                "msg": {
                    "type": "string",
                    "example": "业务错误"
                }
            }
        },
        "response.MessageResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "msg": {
                    "type": "string",
                    "example": "success"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "传入 \"Bearer {access_token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}